% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selectFinal.R
\name{select}
\alias{select}
\title{Implements a Genetic Algorithm for Variable Selection}
\usage{
select(dataset, criterion, model, generation, size)
}
\arguments{
\item{dataset}{dataset that contains the predictors and the response variables. Note: The first column must be the response variable.}

\item{criterion}{function that specifies which objective criterion/fitness function will be used to calculate fitness scores. Note: The default is AIC}

\item{model}{model that user wishes to fit to the data either using \code{lm} or \code{glm}. Note: The default is \code{lm}}

\item{generation}{total number of generations that the user wants through iterate through. Note: The default is 20 generations}

\item{size}{population size Note: The number entered must be even. The default is set to 50 chromosomes}
}
\value{
\code{select} returns a list the best fitted predictors, fitness score, and coefficient values for the predictors.
}
\description{
This R package implements a genetic algorithm for variable selection in regression problems. The functionality of this package extends to linear regressions or generalized linear models.

\code{select} returns a list that includes the best fitted predictors, fitness score, and coefficient values for the predictors.
}
\details{
The goal of the package is to utilize the following components of genetic algortihms:
selection, crossover and mutation in order to improve feature selection.
This type of algorithm mimics the process of natural selection, hence it favors the fittest individuals over all the generations.

This package utilizes crossover and mutation processes adapted to a problem of variable selection. In this package, the population is represented by a matrix and the length (of chromosomes in traditional genetic algorithms) is equivalent to the number of columns of this population matrix. Preceeding the three building blocks we have mentioned, the algorithm requires two stages: initialization and fitness assignment.
}
\examples{
select(dataset = mtcars, criterion = AIC, model = lm, generation = 50, size = 10)
select(dataset = iris, criterion = BIC, model = glm)
select(iris)

}
